+++ tic-tac-toe-challenge/src/App.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 import { GameProvider, useGame } from './context/GameContext';
-import Board from './components/Board';
-import GameModal from './components/GameModal';
+import Board from './components/game/Board';
+import GameModal from './components/game/GameModal';
 import './styles/App.css';
-import Player from './components/Player';
+import Player from './components/game/Player';
 import { formatTime } from './utils/utils'
-import StatsLeft from './components/StatsLeft';
-import StatsRight from './components/StatsRight';
-import GridSelector from './components/GridSelector';
+import GridSelector from './components/game/GridSelector';
+import Stats from './components/stats/Stats';
 
 export const GameLayout: React.FC = () => {
   const {
@@ -28,8 +27,6 @@ export const GameLayout: React.FC = () => {
         <div className="players-board-container d-flex flex-md-row">
           <Player label="PLAYER 1" wins={xWins} className="order-1"/>
           <div className="game-board order-0 order-md-1">
-
-            
             <Board />
           </div>
           <Player label="PLAYER 2" wins={oWins} className="order-2"/>
@@ -40,14 +37,7 @@ export const GameLayout: React.FC = () => {
           </h2>
           <h2>Current Player: {currentPlayer}</h2>
         </div>
-        <div className="row stats-and-right-container">
-          <div className='col-md-6 stats-container'>
-            <StatsLeft />
-          </div>
-          <div className="col-md-6 right-content">
-            <StatsRight />
-          </div>
-        </div>
+        <Stats />
         <h2 className="mt-4">Total Play Time: {formatTime(totalPlayTime)}</h2>
       </div>
       <GameModal />
