--- tic-tac-toe-challenge/src/Board.test.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import '@testing-library/jest-dom';
-import Board from './components/Board';
-
-test('renders the board', () => {
-  // const { getByTestId } = render(<Board />);
-  // // eslint-disable-next-line testing-library/prefer-screen-queries
-  // const squareElement = getByTestId('square');
-  // expect(squareElement).toBeInTheDocument();
-});
