+++ tic-tac-toe-challenge/src/components/game/GridSelector.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useGame } from "../context/GameContext";
+import { useGame } from "../../context/GameContext";
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 function GridSelector() {
