+++ tic-tac-toe-challenge/src/components/stats/StatsLeft.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useGame } from '../context/GameContext';
+import { useGame } from '../../context/GameContext';
 import PlayerStats from './PlayerStats';
 
 function StatsLeft() {
