+++ tic-tac-toe-challenge/src/components/stats/StatsRight.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useGame } from '../context/GameContext';
+import { useGame } from '../../context/GameContext';
 import PlayedGames from './PlayedGames';
 import GamesHistory from './GamesHistory';
